Extract code for working with file system in cloned repository
