+++ Sources/Application/Kysect.Zeya.GitIntegration.Abstraction/ClonedRepository.cs
@@ -1,40 +1,14 @@
-﻿using Kysect.CommonLib.FileSystem;
-using System.IO.Abstractions;
+﻿using System.IO.Abstractions;
 
 namespace Kysect.Zeya.GitIntegration.Abstraction;
 
 public class ClonedRepository(string repositoryRootPath, IFileSystem fileSystem) : IClonedRepository
 {
+    public LocalRepositoryFileSystem FileSystem => new LocalRepositoryFileSystem(repositoryRootPath, fileSystem);
+
     public virtual string GetRepositoryName()
     {
         IDirectoryInfo directoryInfo = fileSystem.DirectoryInfo.New(repositoryRootPath);
         return directoryInfo.Name;
-    }
-
-    public string GetFullPath()
-    {
-        return repositoryRootPath;
-    }
-
-    public bool Exists(string partialPath)
-    {
-        return fileSystem.File.Exists(GetFullPathToFile(partialPath));
-    }
-
-    public string ReadAllText(string partialPath)
-    {
-        return fileSystem.File.ReadAllText(GetFullPathToFile(partialPath));
-    }
-
-    public void WriteAllText(string partialPath, string content)
-    {
-        string fullPathToFile = GetFullPathToFile(partialPath);
-        DirectoryExtensions.EnsureParentDirectoryExists(fileSystem, fullPathToFile);
-        fileSystem.File.WriteAllText(fullPathToFile, content);
-    }
-
-    private string GetFullPathToFile(string partialPath)
-    {
-        return fileSystem.Path.Combine(GetFullPath(), partialPath);
     }
 }
