+++ Sources/Application/Kysect.Zeya.GitIntegration.Abstraction/IClonedRepository.cs
@@ -4,8 +4,5 @@ public interface IClonedRepository
 {
     string GetRepositoryName();
 
-    string GetFullPath();
-    bool Exists(string partialPath);
-    string ReadAllText(string partialPath);
-    void WriteAllText(string partialPath, string content);
+    LocalRepositoryFileSystem FileSystem { get; }
 }
