+++ Sources/Application/Kysect.Zeya.GithubIntegration.Abstraction/ClonedGithubRepository.cs
@@ -4,19 +4,18 @@ using System.IO.Abstractions;
 namespace Kysect.Zeya.GithubIntegration.Abstraction;
 
 public class ClonedGithubRepository(GithubRepositoryName githubMetadata, string repositoryRootPath, IFileSystem fileSystem)
-    : ClonedRepository(repositoryRootPath, fileSystem)
+    : IClonedRepository
 {
-    private readonly IFileSystem _fileSystem = fileSystem;
-
     public GithubRepositoryName GithubMetadata { get; } = githubMetadata;
+    public LocalRepositoryFileSystem FileSystem => new LocalRepositoryFileSystem(repositoryRootPath, fileSystem);
 
-    public override string GetRepositoryName()
+    public string GetRepositoryName()
     {
         return GithubMetadata.FullName;
     }
 
     public string GetWorkflowPath(string workflowName)
     {
-        return _fileSystem.Path.Combine(".github", "workflows", workflowName);
+        return fileSystem.Path.Combine(".github", "workflows", workflowName);
     }
 }
