+++ Sources/Application/Kysect.Zeya.ValidationRules/Rules/Nuget/NugetMetadataHaveCorrectValueValidationRule.cs
@@ -28,7 +28,7 @@ public class NugetMetadataHaveCorrectValueValidationRule(RepositorySolutionAcces
         RepositoryValidationContext repositoryValidationContext = context.GetValidationContext();
         RepositorySolutionAccessor repositorySolutionAccessor = repositorySolutionAccessorFactory.Create(repositoryValidationContext.Repository);
 
-        if (!repositoryValidationContext.Repository.Exists(repositorySolutionAccessor.GetDirectoryBuildPropsPath()))
+        if (!repositoryValidationContext.Repository.FileSystem.Exists(repositorySolutionAccessor.GetDirectoryBuildPropsPath()))
         {
             repositoryValidationContext.DiagnosticCollector.AddDiagnostic(
                 request.DiagnosticCode,
