+++ Sources/Application/Kysect.Zeya.ValidationRules/Rules/SourceCode/ArtifactsOutputEnabledValidationRule.cs
@@ -28,7 +28,7 @@ public class ArtifactsOutputEnabledValidationRule(RepositorySolutionAccessorFact
         RepositoryValidationContext repositoryValidationContext = context.GetValidationContext();
         RepositorySolutionAccessor repositorySolutionAccessor = repositorySolutionAccessorFactory.Create(repositoryValidationContext.Repository);
 
-        if (!repositoryValidationContext.Repository.Exists(repositorySolutionAccessor.GetDirectoryBuildPropsPath()))
+        if (!repositoryValidationContext.Repository.FileSystem.Exists(repositorySolutionAccessor.GetDirectoryBuildPropsPath()))
         {
             repositoryValidationContext.DiagnosticCollector.AddDiagnostic(
                 request.DiagnosticCode,
