+++ Sources/Application/Kysect.Zeya.ValidationRules/Rules/SourceCode/RequiredPackagesAddedValidationRule.cs
@@ -27,7 +27,7 @@ public class RequiredPackagesAddedValidationRule(RepositorySolutionAccessorFacto
         var repositoryValidationContext = context.GetValidationContext();
         RepositorySolutionAccessor repositorySolutionAccessor = repositorySolutionAccessorFactory.Create(repositoryValidationContext.Repository);
 
-        if (!repositoryValidationContext.Repository.Exists(repositorySolutionAccessor.GetDirectoryBuildPropsPath()))
+        if (!repositoryValidationContext.Repository.FileSystem.Exists(repositorySolutionAccessor.GetDirectoryBuildPropsPath()))
         {
             repositoryValidationContext.DiagnosticCollector.AddDiagnostic(
                 request.DiagnosticCode,
